helper.py
import logging
import mimetypes

try:
    import magic
except ImportError:
    magic = None  # If python-magic isn't available, we'll fallback to 'mimetypes'

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

def decrypt_file(encrypted_data: bytes, encryption_key: str) -> bytes:
    """
    Decrypt the file using the given encryption key.
    Implement actual decryption logic here.
    If no encryption is actually needed, just return the input bytes.
    """
    # Placeholder no-op: Just return the encrypted_data as is.
    # In a real scenario, perform actual decryption using encryption_key.
    return encrypted_data

def get_mime_type(file_path: str) -> str:
    """
    Detect the MIME type of a file.
    Uses python-magic for content-based detection if available,
    otherwise falls back to mimetypes based on the file extension.
    """
    if magic:
        try:
            mime = magic.Magic(mime=True)
            return mime.from_file(file_path)
        except Exception:
            pass
    # Fallback to mimetypes
    mime_type, _ = mimetypes.guess_type(file_path)
    return mime_type or "application/octet-stream"