requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiomysql==0.2.0
aiosignal==1.3.1
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.56.1
phonenumbers==8.13.51
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.19
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
yarl==1.18.3
